+++ CHANGELOG.md
@@ -18,8 +18,9 @@ Try to use the following format:
 
 ### Changed
 
-- cg workflow mip-dna -pb option can be given when executing full workflow
+- cg workflow mip-dna -pb and --dry-run options can be set when executing full workflow
 - Changed logic for how case links are retrieved in order to support linking of very old cases
+- Analysis not submitted to Trailblazer if executing MIP workflow in dry-run